Adiciona opção de tv ao sample
