+++ sample/src/main/AndroidManifest.xml
@@ -18,5 +18,6 @@
             </intent-filter>
         </activity>
 
+        <activity android:name=".ActivityMobile" />
     </application>
 </manifest>
