+++ sample/src/main/res/layout-land/activity_mobile.xml
